Revert "Revert "Lazy load Sandpack alone""
This reverts commit 71be160b9c05055f771bc69564e9bdab288da7ba.
