+++ beta/src/components/MDX/Sandpack/index.tsx
@@ -1,144 +1,25 @@
-/*
- * Copyright (c) Facebook, Inc. and its affiliates.
- */
+import * as React from 'react';
+const Sandpack = React.lazy(
+  () =>
+    import('./SandpackWrapper').then(
+      (x) =>
+        new Promise((resolve) => {
+          setTimeout(() => resolve(x), 3000);
+        })
+    ) as any
+);
 
-import React from 'react';
-import {
-  SandpackProvider,
-  SandpackSetup,
-  SandpackFile,
-} from '@codesandbox/sandpack-react';
-
-import {CustomPreset} from './CustomPreset';
-
-type SandpackProps = {
-  children: React.ReactChildren;
-  autorun?: boolean;
-  setup?: SandpackSetup;
-  showDevTools?: boolean;
-};
-
-const sandboxStyle = `
-* {
-  box-sizing: border-box;
-}
-
-body {
-  font-family: sans-serif;
-  margin: 20px;
-  padding: 0;
-}
-
-h1 {
-  margin-top: 0;
-  font-size: 22px;
-}
-
-h2 {
-  margin-top: 0;
-  font-size: 20px;
-}
-
-h3 {
-  margin-top: 0;
-  font-size: 18px;
-}
-
-h4 {
-  margin-top: 0;
-  font-size: 16px;
-}
-
-h5 {
-  margin-top: 0;
-  font-size: 14px;
-}
-
-h6 {
-  margin-top: 0;
-  font-size: 12px;
-}
-
-ul {
-  padding-left: 20px;
-}
-`.trim();
-
-function Sandpack(props: SandpackProps) {
-  let {children, setup, autorun = true, showDevTools = false} = props;
-  const [devToolsLoaded, setDevToolsLoaded] = React.useState(false);
-  let codeSnippets = React.Children.toArray(children) as React.ReactElement[];
-  let isSingleFile = true;
-
-  const files = codeSnippets.reduce(
-    (result: Record<string, SandpackFile>, codeSnippet: React.ReactElement) => {
-      if (codeSnippet.props.mdxType !== 'pre') {
-        return result;
-      }
-      const {props} = codeSnippet.props.children;
-      let filePath; // path in the folder structure
-      let fileHidden = false; // if the file is available as a tab
-      let fileActive = false; // if the file tab is shown by default
-
-      if (props.metastring) {
-        const [name, ...params] = props.metastring.split(' ');
-        filePath = '/' + name;
-        if (params.includes('hidden')) {
-          fileHidden = true;
-        }
-        if (params.includes('active')) {
-          fileActive = true;
-        }
-        isSingleFile = false;
-      } else {
-        if (props.className === 'language-js') {
-          filePath = '/App.js';
-        } else if (props.className === 'language-css') {
-          filePath = '/styles.css';
-        } else {
-          throw new Error(
-            `Code block is missing a filename: ${props.children}`
-          );
-        }
-      }
-      if (result[filePath]) {
-        throw new Error(
-          `File ${filePath} was defined multiple times. Each file snippet should have a unique path name`
-        );
-      }
-      result[filePath] = {
-        code: props.children as string,
-        hidden: fileHidden,
-        active: fileActive,
-      };
-
-      return result;
-    },
-    {}
-  );
-
-  files['/styles.css'] = {
-    code: [sandboxStyle, files['/styles.css']?.code ?? ''].join('\n\n'),
-    hidden: true,
-  };
-
+export default React.memo(function SandpackWrapper(props: any): any {
   return (
-    <div className="sandpack-container my-8" translate="no">
-      <SandpackProvider
-        template="react"
-        customSetup={{...setup, files: files}}
-        autorun={autorun}>
-        <CustomPreset
-          isSingleFile={isSingleFile}
-          showDevTools={showDevTools}
-          onDevToolsLoad={() => setDevToolsLoaded(true)}
-          devToolsLoaded={devToolsLoaded}
-        />
-      </SandpackProvider>
-    </div>
+    <React.Suspense
+      fallback={
+        <pre className="rounded-lg leading-6 h-full w-full overflow-x-auto flex items-center bg-wash dark:bg-gray-95 shadow-lg my-8 text-[13.6px] overflow-hidden">
+          <div className="py-6 pl-5 transition-all duration-1000 font-normal">
+            {props.children}
+          </div>
+        </pre>
+      }>
+      <Sandpack {...props} />
+    </React.Suspense>
   );
-}
-
-Sandpack.displayName = 'Sandpack';
-
-export default Sandpack;
+});
